+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Each release can have sections: "Added", "Changed", "Deprecated", "Removed", "Fixed" and "Security".
 
+## pre-release
+
+### Added
+
+- `AS_BANDS` option for Sentinel-1 orbit, to mosaic ascending and descending bands as different bands [72b396b](https://github.com/gbelouze/geefetch/commit/72b396b22f84051839b3a584d59b8ff1e95e5fd5)
+
 ## [0.5.0](https://github.com/gbelouze/geefetch/compare/v0.5.0...v0.4.2) (2025-02-13)
 
 ### Added