+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Each release can have sections: "Added", "Changed", "Deprecated", "Removed", "Fixed" and "Security".
 
+## pre-release
+
+### Added
+
+- Suppert for [NASADEM](https://developers.google.com/earth-engine/datasets/catalog/NASA_NASADEM_HGT_001) [0d814d5](https://github.com/gbelouze/geefetch/commit/0d814d5b6c99578441356ef6e92ef6b15f055b96)
+
 ## [0.5.0](https://github.com/gbelouze/geefetch/compare/v0.5.0...v0.4.2) (2025-02-13)
 
 ### Added